Refactoring File Structure
